// src/shared/api/models/postApiMakeCardPayment200.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiMakeCardPayment200Data } from './postApiMakeCardPayment200Data';

export type PostApiMakeCardPayment200 = {
    success?: boolean;
    message?: string;
    data?: PostApiMakeCardPayment200Data;
};