// src/shared/api/models/postApiMakeCardPayment400.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiMakeCardPayment400Error } from './postApiMakeCardPayment400Error';

export type PostApiMakeCardPayment400 = {
    statusCode: number;
    error: PostApiMakeCardPayment400Error;
    message: string;
};