// src/shared/api/models/postApiMakeCardPayment500.ts
/**
 * Generated by orval v7.13.2 🍺
 * Do not edit manually.
 * API
 * OpenAPI spec version: 1.0.0
 */
import type { PostApiMakeCardPayment500Error } from './postApiMakeCardPayment500Error';

export type PostApiMakeCardPayment500 = {
    statusCode: number;
    error: PostApiMakeCardPayment500Error;
    message: string;
};